Preserve ABI whether the module signing support is enabled or not
When disabling module signing support, the `sig_ok` field of
`struct module` is dropped. This causes the CRC in `Modules.symvers`
for each symbol accepting a `struct module` to change, since
`genksyms` derives the CRC by recursively checking the definition
of each parameter.

This in turn means that when disabling modules signing support the
debian package build fails because the CRC checked by the ABI step
in `buildcheck.py` won't match the reference.

By making the inclusion of the `sig_ok` symbol unconditional the ABI
is now preserved.

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>