Bump abiname as we don't match the Debian kernel ABI
The ABI of the Apertis kernel packages does not match the one of the Debian
ones due to different configuration options.

In particular, disabling module signing changes the ABI signature of all the
symbols accepting a `struct module` parameter, for which we had a workaround in
https://gitlab.apertis.org/pkg/target/linux/commit/df5b937662ec9fb0

However on `armhf` we also disable Xen support (thus dropping all the
Xen-related exported symbols) and we enable modules that export symbols which
are not tracked by the Debian ABI reference, causing build failures.

We now simply recognize that the two Apertis and Debian ABI do not really match
(binary kernel module compatibility is not a goal at the moment) and thus we
bump the ABI name in `debian/config/defines`.

After that, generated files are updated with:

    $ fakeroot debian/rules debian/control-real

See more details at:
https://kernel-team.pages.debian.net/kernel-handbook/ch-versions.html#s-abi-maintenance

Signed-off-by: Andrew Lee (李健秋) <andrew.lee@collabora.co.uk>