Revert abiname bump
This reverts commit 8e1482b42c700a73f2d0a502bf80cae0db72ea19
"Bump abiname as we don't match the Debian kernel ABI" that was
introduced to avoid complaints from the ABI checker since the Apertis
kernel module ABI does not match the Debian one.

While most of the files got updated with the abiname bump, the
maintscripts have been left behind.

For this reason, the generated binaries are not shipping any
maintscript, causing all kind of breakage: for instance, the
`linux-image` packages misses the postinst script that trigger the
generation of the initrd, breaking the boot process.

Rather than introducing even more delta compared to the upstream
packaging, a later commit will simply disable the ABI check.