Manually refresh single patch
apparmor/0001-apparmor-patch-to-provide-compatibility-with-v2.x-ne.patch

All our patches are maintained through `gbp pq`. But after the import of
Linux version 4.19.98-1 from Debian, `gbp pq` would not cleanly import
the patches.

`quilt push -a` however applies fine

The package can be pushed directly to `dpkg-source` but some of the
patches have fuzz and `dpkg-source` comlains and aborts about it. To
keep the fuzz changes minimal, only refresh the single patch.

Signed-off-by: Ritesh Raj Sarraf <ritesh.sarraf@collabora.com>