Convert package format to that used by gbp pq
The Debian kernel has moved on from v5.4 to v5.5. We are going to continue
using v5.4, however as kernel updates will not be being provided by Debian
we will be pulling updates from the linux stable branch.

Our workflow is going to be:

- Ensure a gbp pq patch branch exists for the Apertis release we're
  targetting
- Fetch latest v5.4.y stable upstream branch in a separate repo
- Add changelog entry for newest v5.4.y kernel version
- Run `debian/bin/genorig.py <path to upstream repo>` to generate
  debianised kernel source tarball
- Run `gbp import-orig --upstream-branch=upstream/linux-5.4.y \
    --debian-branch=apertis/v2021dev2 <path to orig tarball>` to import
  the updated source tarball
- Use `gbp pq rebase` to rebase the Debian patches on the new kernel
  version
- Export updated patch series with `gbp pq export`

As `gbp pq export` changes all the patches to match it's preferred patch
format, handling the large number of patch formatting changes becomes a
time sink. As we no longer expect to be using Debian as an upstream for
this package, we don't need to stay in sync with the patch format to
enable syncing of changes from Debian, so shift all the patches into gbp's
preferred format here so that imports from the stable branch become easier.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>