gitlab-ci: Drop the special bits now folded in main pipeline
We customized some steps from the main pipeline to ensure that
the `debian/control` file gets re-generated on the fly to reflect our
customizations like disabling some build flavors.

The customization has been problematic as it can go out of sync as it
indeed did with the recent changes in this repository, yielding errors
due to the sequence having changed, stages no longer matching, and
artifacts being saved in different locations.

With infrastructure/ci-package-builder!55 the special bits got folded
in the main pipeline, so we can drop the customization here.

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>